requirements.txt
scipy
matplotlib
seaborn
versioneer
pytest
python-rle
